fix comments.
