+++ flink-table/flink-table-planner-blink/src/test/scala/org/apache/flink/table/planner/plan/batch/sql/agg/DistinctAggregateTest.scala
@@ -18,104 +18,8 @@
 
 package org.apache.flink.table.planner.plan.batch.sql.agg
 
-import org.apache.flink.api.scala._
-import org.apache.flink.table.api.scala._
-import org.apache.flink.table.planner.utils.TableTestBase
+import org.apache.flink.table.planner.plan.common.DistinctAggregateTestBase
 
-import org.junit.Test
-
-class DistinctAggregateTest extends TableTestBase {
-  private val util = batchTestUtil()
-  util.addTableSource[(Int, Long, String)]("MyTable", 'a, 'b, 'c)
-  util.addTableSource[(Int, Long, String, String, String)]("MyTable2", 'a, 'b, 'c, 'd, 'e)
-
-  @Test
-  def testSingleDistinctAggregate(): Unit = {
-    util.verifyPlan("SELECT COUNT(DISTINCT a) FROM MyTable")
-  }
-
-  @Test
-  def testMultiDistinctAggregateOnSameColumn(): Unit = {
-    util.verifyPlan("SELECT COUNT(DISTINCT a), SUM(DISTINCT a), MAX(DISTINCT a) FROM MyTable")
-  }
-
-  @Test
-  def testSingleDistinctAggregateAndOneOrMultiNonDistinctAggregate(): Unit = {
-    // case 0x00: DISTINCT on COUNT and Non-DISTINCT on others
-    util.verifyPlan("SELECT COUNT(DISTINCT a), SUM(b) FROM MyTable")
-  }
-
-  @Test
-  def testSingleDistinctAggregateAndOneOrMultiNonDistinctAggregate2(): Unit = {
-    // case 0x01: Non-DISTINCT on COUNT and DISTINCT on others
-    util.verifyPlan("SELECT COUNT(a), SUM(DISTINCT b) FROM MyTable")
-  }
-
-  @Test
-  def testMultiDistinctAggregateOnDifferentColumn(): Unit = {
-    util.verifyPlan("SELECT COUNT(DISTINCT a), SUM(DISTINCT b) FROM MyTable")
-  }
-
-  @Test
-  def testMultiDistinctAndNonDistinctAggregateOnDifferentColumn(): Unit = {
-    util.verifyPlan("SELECT COUNT(DISTINCT a), SUM(DISTINCT b), COUNT(c) FROM MyTable")
-  }
-
-  @Test
-  def testSingleDistinctAggregateWithGrouping(): Unit = {
-    util.verifyPlan("SELECT a, COUNT(a), SUM(DISTINCT b) FROM MyTable GROUP BY a")
-  }
-
-  @Test
-  def testSingleDistinctAggregateWithGroupingAndCountStar(): Unit = {
-    util.verifyPlan("SELECT a, COUNT(*), SUM(DISTINCT b) FROM MyTable GROUP BY a")
-  }
-
-  @Test
-  def testTwoDistinctAggregateWithGroupingAndCountStar(): Unit = {
-    val sqlQuery = "SELECT a, COUNT(*), SUM(DISTINCT b), COUNT(DISTINCT b) FROM MyTable GROUP BY a"
-    util.verifyPlan(sqlQuery)
-  }
-
-  @Test
-  def testTwoDifferentDistinctAggregateWithGroupingAndCountStar(): Unit = {
-    val sqlQuery = "SELECT a, COUNT(*), SUM(DISTINCT b), COUNT(DISTINCT c) FROM MyTable GROUP BY a"
-    util.verifyPlan(sqlQuery)
-  }
-
-  @Test
-  def testSingleDistinctWithFilter(): Unit = {
-    val sqlQuery = "SELECT d, COUNT(DISTINCT c) FILTER (WHERE a > 0) FROM MyTable2 GROUP BY d"
-    util.verifyPlan(sqlQuery)
-  }
-
-  @Test
-  def testMultiDistinctOnSameColumnWithFilter(): Unit = {
-    val sqlQuery = "SELECT d, COUNT(DISTINCT c), COUNT(DISTINCT c) FILTER (WHERE a > 10),\n" +
-      "COUNT(DISTINCT c) FILTER (WHERE a < 10) FROM MyTable2 GROUP BY d"
-    util.verifyPlan(sqlQuery)
-  }
-
-  @Test
-  def TestMultiDistinctOnDifferentColumnWithFilter(): Unit = {
-    val sqlQuery = "SELECT d, COUNT(DISTINCT c), COUNT(DISTINCT c) FILTER (WHERE a > 0),\n" +
-      "COUNT(DISTINCT b) FILTER (WHERE b > 1) FROM MyTable2 GROUP BY d"
-    util.verifyPlan(sqlQuery)
-  }
-
-  @Test
-  def TestMultiDistinctWithFilterAndNonDistinctAgg(): Unit = {
-    val sqlQuery = "SELECT d, COUNT(DISTINCT c), COUNT(DISTINCT c) FILTER (WHERE a > 0),\n" +
-      "MAX(e), MIN(e) FROM MyTable2 GROUP BY d"
-    util.verifyPlan(sqlQuery)
-  }
-
-  @Test
-  def testMultiDistinctAndNonDistinctAggWithFilter(): Unit = {
-    val sqlQuery = "SELECT d, MAX(e), MAX(e) FILTER (WHERE a < 10), COUNT(DISTINCT c),\n" +
-      "COUNT(DISTINCT c) FILTER (WHERE a > 5), COUNT(DISTINCT b) FILTER (WHERE b > 3)\n" +
-      "FROM MyTable2 GROUP BY d"
-    util.verifyPlan(sqlQuery)
-  }
+class DistinctAggregateTest extends DistinctAggregateTestBase {
 
 }
