+++ flink-table/flink-table-planner-blink/src/test/scala/org/apache/flink/table/planner/runtime/batch/sql/agg/DistinctAggregateITCaseBase.scala
@@ -24,7 +24,6 @@ import org.apache.flink.table.planner.runtime.utils.BatchTestBase
 import org.apache.flink.table.planner.runtime.utils.BatchTestBase.row
 import org.apache.flink.table.planner.runtime.utils.TestData._
 import org.apache.flink.types.Row
-
 import org.junit.{Before, Ignore, Test}
 
 import scala.collection.Seq
@@ -316,6 +315,28 @@ abstract class DistinctAggregateITCaseBase extends BatchTestBase {
       "COUNT(DISTINCT a) FILTER (WHERE c > 5), COUNT(DISTINCT b) FILTER (WHERE b > 3)\n" +
       "FROM Table5 GROUP BY e",
       Seq(row(1, 10, 8, 4, 2, 3), row(2, 14, 6, 4, 2, 6), row (3, 12, 5, 2, 1, 3)))
+  }
+
+  @Test
+  def TestDistinctWithFilterWithoutGroupBy(): Unit = {
+    // single distinct agg with filter.
+    checkResult("SELECT COUNT(DISTINCT a) FILTER (WHERE c > 0) FROM Table5",
+      Seq(row(4)))
+
+    // multi distinct aggs on same column with filter.
+    checkResult("SELECT COUNT(DISTINCT a), COUNT(DISTINCT a) FILTER (WHERE c > 10),\n" +
+      "COUNT(DISTINCT a) FILTER (WHERE c < 10) FROM Table5",
+      Seq(row(5, 1, 4)))
+
+    // multi distinct aggs on different columns with filter.
+    checkResult("SELECT COUNT(DISTINCT a), COUNT(DISTINCT a) FILTER (WHERE c > 0),\n" +
+      "COUNT(DISTINCT b) FILTER (WHERE b > 1) FROM Table5",
+      Seq(row(5, 4, 14)))
+
+    // multi distinct aggs with non-distinct agg with filter.
+    checkResult("SELECT MAX(e), MAX(e) FILTER (WHERE c < 10), COUNT(DISTINCT a),\n" +
+      "COUNT(DISTINCT a) FILTER (WHERE c > 5), COUNT(DISTINCT b) FILTER (WHERE b > 3) FROM Table5",
+      Seq(row(3, 3, 5, 2, 12)))
   }
 
   // TODO remove Ignore after supporting generated code cloud be splitted into
